Goed

* commit '0b1c708a2ea24b4f9163d97d874767adad5ec301':
  update member variable from merge conflict
  update core machine  constructor to initialise current config to 0
  switching products
  app houdt eerste en laatste machine per product bij, voor bijsturing productie; buffer heeft ook zijn next machines, zodat deze weet wanneer hij een laatste buffer is; application heeft een aantal functies erbij voor switchen van product (tijdelijk nu

+++ src/libs/models/MachineConfiguration.cpp
@@ -81,8 +81,8 @@ uint16_t MachineConfiguration::getInitializationDurationInSeconds() const {
   return initializationDurationInSeconds;
 }
 
-uint16_t MachineConfiguration::getInitializationDurationInMilliseconds() const {
-  return getInitializationDurationInSeconds() * 1000;
+uint32_t MachineConfiguration::getInitializationDurationInMilliseconds() const {
+  return static_cast<uint32_t>(getInitializationDurationInSeconds() * 1000);
 }
 
 uint16_t MachineConfiguration::getOutputBufferSize() const {
